refactor: EventSection component
